Make Ceph collector to pick up cluster name
Previously the prefix on admin socket names
was a config setting, had to be manually set
for one and only one cluster name.  In the
output stats, the cluster name was not used,
'ceph' was always passed through.

This change makes the collector gather stats
for all clusters it sees, and include the
name of the cluster in the metric names.